+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/Krypton Toast/Controls Visuals/Basic/LTR/VisualToastBasicForm.cs
@@ -12,7 +12,7 @@ using Timer = System.Windows.Forms.Timer;
 
 namespace Krypton.Toolkit.Utilities;
 
-internal partial class VisualToastBasicForm : KryptonForm
+internal partial class VisualToastBasicForm : VisualToastBaseForm
 {
     #region Instance Fields
 
@@ -139,41 +139,24 @@ internal partial class VisualToastBasicForm : KryptonForm
     }
 
     private void UpdateLocation() =>
-        //Once loaded, position the form, or position it to the bottom left of the screen with added padding
-        Location = _basicToastNotificationData.NotificationLocation ?? new Point(Screen.PrimaryScreen!.WorkingArea.Width - Width - 5,
-            Screen.PrimaryScreen.WorkingArea.Height - Height - 5);
+        // Once loaded, position the form, or default to bottom-right with DPI-scaled edge padding.
+        Location = _basicToastNotificationData.NotificationLocation ?? GetDefaultBottomRightLocation();
 
     private void ReportToastLocation() => klblToastLocation.Text = _basicToastNotificationData.ReportToastLocation ? $"Location: X: {Location.X}, Y: {Location.Y}" : string.Empty;
 
     private void VisualToastNotificationBasicForm_Load(object sender, EventArgs e)
     {
-        UpdateSizing();
+        ShowCloseButton();
+
+        ApplyToastDpiLayout();
 
         UpdateLocation();
 
         ReportToastLocation();
 
-        ShowCloseButton();
-
         _timer?.Start();
 
         _soundPlayer?.Play();
-    }
-
-    private void UpdateSizing()
-    {
-        if (FormBorderStyle == FormBorderStyle.None)
-        {
-            // Add some height, if form border style equals 'None'
-
-            var width = Size.Width;
-
-            // ToDo: Use scaling here, to support larger screens
-
-            var height = Size.Height + SharedStaticConstants.DEFAULT_PADDING;
-
-            Size = new Size(width, height);
-        }
     }
 
     private void VisualToastNotificationBasicForm_Resize(object? sender, EventArgs e)
@@ -197,14 +180,8 @@ internal partial class VisualToastBasicForm : KryptonForm
 
     private void kbtnDismiss_Click(object sender, EventArgs e) => Close();
 
-    private void ShowCloseButton()
-    {
-        CloseBox = _basicToastNotificationData.ShowCloseBox ?? false;
-
-        FormBorderStyle = CloseBox ? FormBorderStyle.Fixed3D : FormBorderStyle.FixedSingle;
-
-        ControlBox = _basicToastNotificationData.ShowCloseBox ?? false;
-    }
+    private void ShowCloseButton() =>
+        ApplyCloseBoxChrome(_basicToastNotificationData.ShowCloseBox ?? false);
 
     private void ShowDoNotShowAgainOption()
     {