+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/Krypton Toast/Controls Visuals/User Input/LTR/ProgressBar/VisualToastMaskedTextBoxInputWithProgressBarForm.cs
@@ -74,9 +74,8 @@ internal partial class VisualToastMaskedTextBoxInputWithProgressBarForm : Visual
 
     private void UpdateLocation()
     {
-        //Once loaded, position the form, or position it to the bottom left of the screen with added padding
-        Location = _data.NotificationLocation ?? new Point(Screen.PrimaryScreen!.WorkingArea.Width - Width - 5,
-            Screen.PrimaryScreen.WorkingArea.Height - Height - 5);
+        // Once loaded, position the form, or default to bottom-right with DPI-scaled edge padding.
+        Location = _data.NotificationLocation ?? GetDefaultBottomRightLocation();
     }
 
     private void UpdateIcon()
@@ -102,23 +101,19 @@ internal partial class VisualToastMaskedTextBoxInputWithProgressBarForm : Visual
     {
         UpdateIcon();
 
-        UpdateLocation();
-
         ShowCloseButton();
+
+        ApplyToastDpiLayout();
+
+        UpdateLocation();
 
         kbtnDismiss.Text = KryptonManager.Strings.ToastNotificationStrings.Dismiss;
 
         _timer.Start();
     }
 
-    private void ShowCloseButton()
-    {
-        CloseBox = _data.ShowCloseBox ?? false;
-
-        FormBorderStyle = CloseBox ? FormBorderStyle.Fixed3D : FormBorderStyle.FixedSingle;
-
-        ControlBox = _data.ShowCloseBox ?? false;
-    }
+    private void ShowCloseButton() =>
+        ApplyCloseBoxChrome(_data.ShowCloseBox ?? false);
 
     public new DialogResult ShowDialog()
     {